+++ src/library/module_mgr.h
@@ -33,7 +33,7 @@ struct module_info {
     module_id m_id;
     std::string m_contents;
     // Hash of the Lean source (after normalizing line endings) yielding this module:
-    // - if m_source == LEAN, hash_data(remove_cr(m_contents))
+    // - if m_source == LEAN, hash64_str(remove_cr(m_contents))
     // - if m_source == OLEAN, value loaded from the .olean
     uint64 m_src_hash;
     // Transitive hash of all source code this module was built from,
@@ -71,7 +71,7 @@ struct module_info {
 
     module_info() {}
 
-    module_info(module_id const & id, std::string const & contents, unsigned src_hash, unsigned trans_hash, module_src src)
+    module_info(module_id const & id, std::string const & contents, uint64 src_hash, uint64 trans_hash, module_src src)
             : m_id(id), m_contents(contents), m_src_hash(src_hash), m_trans_hash(trans_hash), m_source(src) {}
 };
